models/model2/requirements.txt
flask
torch
transformers
Pillow
numpy
scikit-image
flask-cors